Implement create and delete options
Unless one of these are given, no action will be performed. When
neither is specified, no update action will be taken; so
`--no-op` is currently redundant, but might become useful again if a
default mode of operation is established.

Monitoring is now orthogonal to the invoked operation, so combining
`--no-wait` and `--no-op` becomes legal.