Use EDNS-free update messages
The way EDNS is currently implemented in `trust-dns`, I do not see a
way to reuse the code for constructing the update messages and
implement TSIG outside of `trust-dns` proper at the same time.

Both eliminating the code duplication introduced by this commit, and
upstreaming at least a part of the upcoming TSIG implementation are
planned, once it is clear what shape the changes changes to
`trust-dns` might take.