query: Improve output, introduce `--fmt`
This introduces a new output format "zone", in addition to the single
one used up to now, now christened "short". By default, if only a
single type is queried, "short" is used. Also, when the types are just
A and AAAA, "short" is used as well. The default can be overridden by
the new `--fmt` (or `-f`) option.